Rename waker module to event
